+++ src/modules/generator/pedersen_impl.h
