+++ src/modules/rangeproof/rangeproof_impl.h
@@ -13,9 +13,9 @@
 #include "../../hash_impl.h"
 #include "../../util.h"
 
-#include "pedersen.h"
-#include "rangeproof.h"
-#include "borromean.h"
+#include "modules/generator/pedersen.h"
+#include "modules/rangeproof/borromean.h"
+#include "modules/rangeproof/rangeproof.h"
 
 SECP256K1_INLINE static void secp256k1_rangeproof_pub_expand(secp256k1_gej *pubs,
  int exp, size_t *rsizes, size_t rings, const secp256k1_ge* genp) {
