+++ src/secp256k1.c
@@ -38,8 +38,6 @@
 
 #ifdef ENABLE_MODULE_RANGEPROOF
 # include "include/secp256k1_rangeproof.h"
-# include "modules/rangeproof/pedersen.h"
-# include "modules/rangeproof/rangeproof.h"
 #endif
 
 #ifdef ENABLE_MODULE_ECDSA_S2C
